+++ extensions/sql-migration/src/wizard/databaseSelectorPage.ts
@@ -11,6 +11,7 @@ import * as constants from '../constants/strings';
 import { debounce } from '../api/utils';
 import * as styles from '../constants/styles';
 import { IconPathHelper } from '../constants/iconPathHelper';
+import { getDatabasesList, excludeDatabses } from '../api/sqlUtils';
 
 export class DatabaseSelectorPage extends MigrationWizardPage {
 	private _view!: azdata.ModelView;
@@ -233,16 +234,10 @@ export class DatabaseSelectorPage extends MigrationWizardPage {
 	}
 
 	private async _loadDatabaseList(stateMachine: MigrationStateModel, selectedDatabases: string[]): Promise<void> {
-		const excludeDbs: string[] = [
-			'master',
-			'tempdb',
-			'msdb',
-			'model'
-		];
+		const allDatabases = (<azdata.DatabaseInfo[]>await getDatabasesList(this.migrationStateModel.sourceConnectionId));
 
-		const databaseList = (<azdata.DatabaseInfo[]>await this.migrationStateModel
-			.getDatabasesList())
-			.filter(database => !excludeDbs.includes(database.options.name))
+		const databaseList = allDatabases
+			.filter(database => !excludeDatabses.includes(database.options.name))
 			|| [];
 
 		databaseList.sort((a, b) => a.options.name.localeCompare(b.options.name));
